fix: save Dish thunk
